requirements.txt
astroid==2.3.3
autopep8==1.4.4
certifi==2019.11.28
cffi==1.13.2
colorama==0.4.1
cryptography==2.8
emoji==0.5.4
future==0.18.2
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pycodestyle==2.5.0
pycparser==2.19
pylint==2.4.4
python-telegram-bot==12.2.0
six==1.13.0
tornado==6.0.3
wrapt==1.11.2
youtube-dl==2019.11.28
